conditional removal of convox.site alb rules
if the service has at least one custom domain and
InternalDomains=No then remove the rules for the
convox.site routing
